chore: rearrange exports
